requirements.txt
requests
pytest
allure-pytest
python-dotenv
selenium
webdriver-manager
pytest-xdist
